adapted SVC subsampling
The number of samples per class in subsampling for classification was set to the number of samples in the smallest class (instead of rounding down to full 10's). This also required a change of the cross-validation function where it was enforced that k cross-validation folds are realized (instead of reducing it in limit cases, e.g. n = 21 & k = 10 -> k = 7; which previously never gave problems when mod(n,10) = 0 & k = 10).